Connect code comments to orgs part 2 (#7205)

+++ pkg/localstore/threads_mock.go
@@ -9,14 +9,14 @@ import (
 )
 
 type MockThreads struct {
-	Get    func(ctx context.Context, id int64) (*sourcegraph.Thread, error)
+	Get    func(ctx context.Context, id int32) (*sourcegraph.Thread, error)
 	Create func(ctx context.Context, newThread *sourcegraph.Thread) (*sourcegraph.Thread, error)
-	Update func(ctx context.Context, id, repoID int64, archived *bool) (*sourcegraph.Thread, error)
+	Update func(ctx context.Context, id, repoID int32, archived *bool) (*sourcegraph.Thread, error)
 }
 
 func (s *MockThreads) MockGet_Return(t *testing.T, returns *sourcegraph.Thread, returnsErr error) (called *bool) {
 	called = new(bool)
-	s.Get = func(ctx context.Context, id int64) (*sourcegraph.Thread, error) {
+	s.Get = func(ctx context.Context, id int32) (*sourcegraph.Thread, error) {
 		*called = true
 		return returns, returnsErr
 	}
@@ -34,7 +34,7 @@ func (s *MockThreads) MockCreate_Return(t *testing.T, returns *sourcegraph.Threa
 
 func (s *MockThreads) MockUpdate_Return(t *testing.T, returns *sourcegraph.Thread, returnsErr error) (called *bool) {
 	called = new(bool)
-	s.Update = func(ctx context.Context, id, repoID int64, archived *bool) (*sourcegraph.Thread, error) {
+	s.Update = func(ctx context.Context, id, repoID int32, archived *bool) (*sourcegraph.Thread, error) {
 		*called = true
 		return returns, returnsErr
 	}
